+++ __tests__/components/App/SortenUrlForm.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import UrlForm from '@/components/App/UrlForm';
+import ShortenUrlForm, { HomeText } from '@/components/App/ShortenUrlForm';
 
 describe('UrlForm component', () => {
     const testUrl = 'https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message';
@@ -10,7 +10,7 @@ describe('UrlForm component', () => {
         const mockOnSubmit = jest.fn();
         const mockClearError = jest.fn();
         render(
-            <UrlForm
+            <ShortenUrlForm
                 url={testUrl}
                 setUrl={mockSetUrl}
                 onSubmit={mockOnSubmit}
@@ -28,7 +28,7 @@ describe('UrlForm component', () => {
         const mockOnSubmit = jest.fn();
         const mockClearError = jest.fn();
         render(
-            <UrlForm
+            <ShortenUrlForm
                 url={testUrl}
                 setUrl={mockSetUrl}
                 onSubmit={mockOnSubmit}
@@ -48,7 +48,7 @@ describe('UrlForm component', () => {
         const mockOnSubmit = jest.fn();
         const mockClearError = jest.fn();
         render(
-            <UrlForm
+            <ShortenUrlForm
                 url={testUrl}
                 setUrl={mockSetUrl}
                 onSubmit={mockOnSubmit}
@@ -65,7 +65,7 @@ describe('UrlForm component', () => {
     it('renders error message if error prop is passed', () => {
         const errorMessage = 'Enter a valid URL';
         render(
-            <UrlForm
+            <ShortenUrlForm
                 url={testUrl}
                 setUrl={jest.fn()}
                 onSubmit={jest.fn()}
@@ -77,5 +77,36 @@ describe('UrlForm component', () => {
         const errorElement = screen.getByText(errorMessage);
         expect(errorElement).toBeInTheDocument();
         expect(errorElement).toHaveTextContent(errorMessage);
+    });
+
+    it('renders the main heading correctly', () => {
+        render(<HomeText />);
+        const mainHeading = screen.getByText('Shorten Your URL');
+        expect(mainHeading).toBeInTheDocument();
+        expect(mainHeading).toHaveClass(
+            'text-3xl md:text-6xl xl:text-7xl sm:text-5xl text-center text-white font-semibold pb-2 lg:pb-4'
+        );
+    });
+
+    it('renders the subheading correctly', () => {
+        render(<HomeText />);
+        const subHeading = screen.getByText('Perfect Links Every Time');
+        expect(subHeading).toBeInTheDocument();
+        expect(subHeading).toHaveClass(
+            'text-2xl sm:text-3xl md:text-4xl xl:text-5xl text-center text-white font-semibold'
+        );
+    });
+
+    it('renders the paragraph text correctly', () => {
+        render(<HomeText />);
+        const paragraph = screen.getByText(/Ready to shorten your URL\? Enter your/i);
+        expect(paragraph).toBeInTheDocument();
+        expect(paragraph).toHaveClass('xl:text-xl text-base text-white mt-4 text-center');
+    });
+
+    it('renders the paragraph text with a line break for small screens', () => {
+        render(<HomeText />);
+        const paragraph = screen.getByText(/Ready to shorten your URL\? Enter your/i);
+        expect(paragraph.innerHTML).toContain('<br class="sm:hidden">');
     });
 });
