+++ __tests__/components/Dashhboard/UrlListItem.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import UrlListItem from '@/components/Dashboard/UrlListItem';
+import UrlListItem, { CopyButton, DeleteButton } from '@/components/Dashboard/UrlListItem';
 import useAuthenticated from '@/hooks/useAuthenticated';
 import { deleteUrlApi } from '@/services/api';
 
@@ -92,5 +92,51 @@ describe('UrlListItem', () => {
         });
 
         alertMock.mockRestore();
+    });
+});
+
+describe('DeleteButton', () => {
+    const onDelete = jest.fn();
+
+    it('renders correctly', () => {
+        render(<DeleteButton isLoading={false} onDelete={onDelete} />);
+        const deleteButton = screen.getByTestId('delete-button');
+        expect(deleteButton).toBeInTheDocument();
+    });
+
+    it('triggers onDelete when clicked', () => {
+        render(<DeleteButton isLoading={false} onDelete={onDelete} />);
+        const deleteButton = screen.getByTestId('delete-button');
+        fireEvent.click(deleteButton);
+        expect(onDelete).toHaveBeenCalledTimes(1);
+    });
+
+    it('displays loader when isLoading is true', () => {
+        render(<DeleteButton isLoading={true} onDelete={onDelete} />);
+        const loader = screen.getByTestId('loader');
+        expect(loader).toBeInTheDocument();
+    });
+
+    it('disables button when isLoading is true', () => {
+        render(<DeleteButton isLoading={true} onDelete={onDelete} />);
+        const deleteButton = screen.getByTestId('delete-button');
+        expect(deleteButton).toBeDisabled();
+    });
+});
+
+describe('CopyButton', () => {
+    const onCopy = jest.fn();
+
+    it('renders correctly', () => {
+        render(<CopyButton onCopy={onCopy} />);
+        const copyButton = screen.getByTestId('copy-button');
+        expect(copyButton).toBeInTheDocument();
+    });
+
+    it('triggers onCopy when clicked', () => {
+        render(<CopyButton onCopy={onCopy} />);
+        const copyButton = screen.getByTestId('copy-button');
+        fireEvent.click(copyButton);
+        expect(onCopy).toHaveBeenCalledTimes(1);
     });
 });
